last version

+++ src/app/back-office/club-list/club-list.component.ts
@@ -14,6 +14,7 @@ export class ClubListComponent implements OnInit {
   faThumbsUp = faThumbsUp;
   faThumbsDown = faThumbsDown;
   chart: any;
+  selectedFiles: { [key: number]: File } = {}; 
  
   // Propriétés pour les statistiques
   totalClubCount: number = 0;
@@ -41,11 +42,13 @@ export class ClubListComponent implements OnInit {
   }
 
   deleteClub(id: number): void {
-    this.clubService.deleteClub(id).subscribe(() => {
-      this.clubs = this.clubs.filter(club => club.id !== id);
-      this.updateStatistics();
-      this.renderChart(); 
-    });
+    if (confirm('Êtes-vous sûr de vouloir supprimer ce club ?')) {
+      this.clubService.deleteClub(id).subscribe(() => {
+        this.clubs = this.clubs.filter(club => club.id !== id);
+        this.updateStatistics();
+        this.renderChart();
+      });
+    }
   }
 
   updateClub(id: number): void {
@@ -190,4 +193,32 @@ export class ClubListComponent implements OnInit {
       });
     }
   }
+
+
+  onFileSelected(event: any, clubId: number): void {
+    const file: File = event.target.files[0];
+    this.selectedFiles[clubId] = file;
+  }
+
+ uploadPhoto(clubId: number): void {
+  const file = this.selectedFiles[clubId];
+  if (file) {
+    const formData = new FormData();
+    formData.append('file', file, file.name);
+
+    this.clubService.uploadPhoto(clubId, formData).subscribe(
+      (response) => {
+        // Mettre à jour l'URL de la photo du club après téléchargement
+        const club = this.clubs.find(club => club.id === clubId);
+        if (club) {
+          club.photo = response.photo; // Assurez-vous que la structure de la réponse correspond à ce que vous attendez
+        }
+      },
+      (error) => {
+        console.error('Error uploading photo', error);
+      }
+    );
+  }
+}
+
   }
